#24: Support for segments over 2 GB

+++ core/src/main/java/com/yandex/yoctodb/util/immutable/impl/FixedLengthByteArraySortedSet.java
@@ -82,7 +82,7 @@ public final class FixedLengthByteArraySortedSet
     public Buffer get(final int i) {
         assert 0 <= i && i < size;
 
-        return elements.slice(i * elementSize, elementSize);
+        return elements.slice(((long) i) * elementSize, elementSize);
     }
 
     @Override
